remove RentalValidationService and its interface
