--- CarRental/CarRental.Logic/Services/IServices/IRentalValidationService.cs
@@ -1,7 +1,0 @@
-﻿using CarRental.Logic.Models;
-
-namespace CarRental.Logic.Services.IServices;
-public interface IRentalValidationService
-{
-    bool IsAllValid(RentalViewModel rental);
-}
