+++ CarRental/CarRental.Logic/Services/RentalService.cs
@@ -4,7 +4,10 @@ using CarRental.DAL.Entities;
 using CarRental.DAL.Repositories;
 using CarRental.Logic.Models;
 using CarRental.Logic.Services.IServices;
+using FluentValidation;
+using FluentValidation.Results;
 using Microsoft.Extensions.Logging;
+using System.Runtime.CompilerServices;
 
 namespace CarRental.Logic.Services;
 
@@ -14,14 +17,16 @@ public class RentalService : IRentalService
     private readonly IRepository<Rental> _rentalRepository;
     private readonly IMapper _mapper;
     private readonly ILogger _logger;
+    private readonly IValidator<RentalViewModel> _validator;
 
     public RentalService(ICarService carService, IMapper mapper, IRepository<Rental> rentalRepository,
-        ILogger<RentalService> logger)
+        ILogger<RentalService> logger, IValidator<RentalViewModel> validator)
     {
         _carService = carService;
         _rentalRepository = rentalRepository;
         _mapper = mapper;
         _logger = logger;
+        _validator = validator;
     }
 
     public List<RentalViewModel> GetAll()
@@ -38,22 +43,24 @@ public class RentalService : IRentalService
 
     public void Create(RentalViewModel model)
     {
-        if (IsRentalValidForCreate(model))
+        if (!(IsRentalValidForCreate(model) && IsAllValid(model)))
         {
-            var rental = _mapper.Map<Rental>(model);
-            _rentalRepository.Insert(rental);
-            _logger.LogInformation($"Rental for car with id {model.CarId} and customer with id {model.CustomerId} was created.");
+            throw new ArgumentException("Rental is not valid for create");
         }
+        var rental = _mapper.Map<Rental>(model);
+        _rentalRepository.Insert(rental);
+        _logger.LogInformation($"Rental for car with id {model.CarId} and customer with id {model.CustomerId} was created.");
     }
 
     public void Update(RentalViewModel model)
     {
-        if (model is not null)
+        if (!(model is not null && IsAllValid(model)))
         {
-            var rental = _mapper.Map<Rental>(model);
-            _logger.LogInformation($"Rental with id {model.Id} was updated.");
-            _rentalRepository.Update(rental);
+            throw new ArgumentException("Rental is not valid for update");
         }
+        var rental = _mapper.Map<Rental>(model);
+        _logger.LogInformation($"Rental with id {model.Id} was updated.");
+        _rentalRepository.Update(rental);
     }
 
     public void Delete(int id)
@@ -153,4 +160,24 @@ public class RentalService : IRentalService
         var rentals = GetAll();
         return rentals.Where(predicate);
     }
+
+    #region Validation
+    public bool IsAllValid(RentalViewModel rental)
+    {
+        var validationResult = _validator.Validate(rental, options =>
+        {
+            options.IncludeAllRuleSets();
+        });
+        LogErrors(validationResult);
+        return validationResult.IsValid;
+    }
+
+    private void LogErrors(ValidationResult validationResult)
+    {
+        foreach (var failure in validationResult.Errors)
+        {
+            _logger.LogInformation("Property " + failure.PropertyName + " failed validation. Error was: " + failure.ErrorMessage);
+        }
+    }
+    #endregion Validation
 }
